ROLES AND PERMISSIONS  - FIRST CHECK-IN =======================================
WARNING:  DEV IS CURRENTLY VERY UNSTABLE.

This is a mega-checkin of the new Roles system.   A lot of changes have
been made in core and modules.

Currently there are a lot of rough edges and known problems.  We are
working hard on these .. .the reason for getting this into HEAD at this
stage is enable us to move faster (our branch was diverging from HEAD
too much).

Please keep an eye on http://docs.moodle.org/en/Roles for current status
and information for developers on how to use the new Roles system.

+++ course/category.php
@@ -289,6 +289,9 @@
         }
 
         foreach ($courses as $acourse) {
+          
+          	$context = get_context_instance(CONTEXT_COURSE, $acourse->id);
+          
             $count++;
             $up = ($count > 1 || !$atfirstpage);
             $down = ($count < $numcourses || !$atlastpage);
@@ -305,24 +308,40 @@
                     echo '<a title="'.$strassignteachers.'" href="'.$CFG->wwwroot.'/course/teacher.php?id='.
                          $acourse->id.'">'.
                          '<img src="'.$CFG->pixpath.'/t/user.gif" height="11" width="11" border="0" alt="'.$strassignteachers.'" /></a> ';
-                    echo '<a title="'.$strdelete.'" href="delete.php?id='.$acourse->id.'">'.
-                         '<img src="'.$CFG->pixpath.'/t/delete.gif" height="11" width="11" border="0" alt="'.$strdelete.'" /></a> ';
-                    if (!empty($acourse->visible)) {
-                        echo '<a title="'.$strhide.'" href="category.php?id='.$category->id.'&amp;page='.$page.
-                             '&amp;perpage='.$perpage.'&amp;hide='.$acourse->id.'&amp;sesskey='.$USER->sesskey.'">'.
-                             '<img src="'.$CFG->pixpath.'/t/hide.gif" height="11" width="11" border="0" alt="'.$strhide.'" /></a> ';
-                    } else {
-                        echo '<a title="'.$strshow.'" href="category.php?id='.$category->id.'&amp;page='.$page.
-                             '&amp;perpage='.$perpage.'&amp;show='.$acourse->id.'&amp;sesskey='.$USER->sesskey.'">'.
-                             '<img src="'.$CFG->pixpath.'/t/show.gif" height="11" width="11" border="0" alt="'.$strshow.'" /></a> ';
+					if (has_capability('moodle/course:delete', $context->id)) {
+                        echo '<a title="'.$strdelete.'" href="delete.php?id='.$acourse->id.'">'.
+                             '<img src="'.$CFG->pixpath.'/t/delete.gif" height="11" width="11" border="0" alt="'.$strdelete.'" /></a> ';
+                    }
+
+					if (has_capability('moodle/course:visibility', $context->id)) {
+                        if (!empty($acourse->visible)) {
+                            echo '<a title="'.$strhide.'" href="category.php?id='.$category->id.'&amp;page='.$page.
+                                '&amp;perpage='.$perpage.'&amp;hide='.$acourse->id.'&amp;sesskey='.$USER->sesskey.'">'.
+                                '<img src="'.$CFG->pixpath.'/t/hide.gif" height="11" width="11" border="0" alt="'.$strhide.'" /></a> ';
+                        } else {
+                            echo '<a title="'.$strshow.'" href="category.php?id='.$category->id.'&amp;page='.$page.
+                                '&amp;perpage='.$perpage.'&amp;show='.$acourse->id.'&amp;sesskey='.$USER->sesskey.'">'.
+                                '<img src="'.$CFG->pixpath.'/t/show.gif" height="11" width="11" border="0" alt="'.$strshow.'" /></a> ';
+                        }
+                    }
+
+					if (has_capability('moodle/site:backup', $context->id)) {
+	                    echo '<a title="'.$strbackup.'" href="../backup/backup.php?id='.$acourse->id.'">'.
+	                         '<img src="'.$CFG->pixpath.'/t/backup.gif" height="11" width="11" border="0" alt="" /></a> ';
+					}
+					
+					if (has_capability('moodle/site:restore', $context->id)) {
+	                    echo '<a title="'.$strrestore.'" href="../files/index.php?id='.$acourse->id.
+	                             '&amp;wdir=/backupdata">'.
+	                             '<img src="'.$CFG->pixpath.'/t/restore.gif" height="11" width="11" border="0" alt="" /></a> ';
                     }
 
                     echo '<a title="'.$strbackup.'" href="../backup/backup.php?id='.$acourse->id.'">'.
                          '<img src="'.$CFG->pixpath.'/t/backup.gif" height="11" width="11" border="0" alt="'.$strbackup.'" /></a> ';
 
-                        echo '<a title="'.$strrestore.'" href="../files/index.php?id='.$acourse->id.
-                             '&amp;wdir=/backupdata">'.
-                             '<img src="'.$CFG->pixpath.'/t/restore.gif" height="11" width="11" border="0" alt="'.$strrestore.'" /></a> ';
+                    echo '<a title="'.$strrestore.'" href="../files/index.php?id='.$acourse->id.
+                         '&amp;wdir=/backupdata">'.
+                         '<img src="'.$CFG->pixpath.'/t/restore.gif" height="11" width="11" border="0" alt="'.$strrestore.'" /></a> ';
 
                     if ($up) {
                         echo '<a title="'.$strmoveup.'" href="category.php?id='.$category->id.'&amp;page='.$page.
@@ -399,14 +418,16 @@
         print_single_button('category.php', $options, get_string('resortcoursesbyname'), 'get');
     }
 
-    if (iscreator()) {         /// Print button to create a new course
+	$context = get_context_instance(CONTEXT_SYSTEM, SITEID);
+    if (has_capability('moodle/course:create', $context->id)) {         /// Print button to create a new course
         unset($options);
         $options['category'] = $category->id;
         print_single_button('edit.php', $options, get_string('addnewcourse'), 'get');
         echo '<br />';
     }
 
-    if (isadmin()) {           /// Print form to rename the category
+	$context = get_context_instance(CONTEXT_COURSECAT, $id);
+    if (has_capability('moodle/category:update', $context->id)) {           /// Print form to rename the category
         $strrename= get_string('rename');
         echo '<form name="renameform" action="category.php" method="post">';
         echo '<input type="hidden" name="id" value="'.$category->id.'" />';
@@ -415,10 +436,10 @@
         echo '<input type="submit" value="'.$strrename.'" />';
         echo "</form>";
         echo "<br />";
-
-        print_course_search();
-
-    }
+	}
+    
+	print_course_search();
+    
     echo "</center>";
 
     print_footer();
